CRUD of book (#34)

+++ src/main/webapp/WEB-INF/views/elements/admin/category.jsp
@@ -33,7 +33,7 @@
 					<td>3</td>
 					<td>100.000</td>
 					<td>
-						<a href="admin/category/${category.MATL}/delete.htm"><button class="btn btn-outline-primary"><i class="bi bi-trash3-fill"></i></button></a>
+						<a href="admin/category/${category.MATL}/delete.htm" onclick="return confirm('Bạn có chắc muốn xóa thể loại ${category.TENTL}?')"><button class="btn btn-outline-primary"><i class="bi bi-trash3-fill"></i></button></a>
 						<a href="admin/category/${category.MATL}/update.htm"><button class="btn btn-outline-primary"><i class="bi bi-pencil-square"></i></button></a>
 					</td>
 				</tr>
